platform/slo-gate/requirements.txt
fastapi==0.116.1
httpx==0.28.1
pydantic==2.11.7
pyyaml==6.0.2
uvicorn[standard]==0.35.0
